Stab at memory layer for int -> long on node,rel,prop ids. Still have to fix IntArray and store layer.

+++ community/kernel/src/main/java/org/neo4j/kernel/impl/core/PropertyIndexManager.java
@@ -195,7 +195,7 @@ public class PropertyIndexManager
         {
             return index;
         }
-        int id = idGenerator.nextId( PropertyIndex.class );
+        int id = (int) idGenerator.nextId( PropertyIndex.class );
         index = new PropertyIndex( key, id );
         hook.addIndex( index );
         persistenceManager.createPropertyIndex( key, id );
